Use the Exit exception to return a finish status in primitive_structure_from_cif
This is preferred over raising a generic exception and catching it
in the CifCleanWorkChain, as in that case the workfunction will be
marked as Excepted whereas really something happened that we foresaw.
A more apt description is to have it marked as Finished but give it
a non-zero finish status to indicate failure.