CifCleanWorkChain: add exception handling for Seekpath symmetry failure
For certain (malformed) cif files, Seekpath will fail to construct a
primitive structure with the message:

    ValueError: Problem creating primitive cell, I found the following
                group of atoms with len != 1: (0, 3)

This will cause the workfunction `primitive_structure_from_cif` to except
and therewith the WorkChain as well.

Here we add specific exception handling for this problem in the workfunction
and add a new exit status for it in the workchain.