Use `--dont-treat-dots-as-underscores` in CifCleanWorkChain
By default `cif_select` in `cod-tools` will treat dots as underscores
and will replace the former with the latter. However, in v0.2.1 it
will not check if this results in duplicate keys, which if the case
will except the `CifData` constructor, causing the workchain to fail.
As a workaround for version of `cod-tools` with this bug, we can pass
the option `--dont-treat-dots-as-underscores` which will stop it from
converting dots to underscores.