Implement new generic command line script to launch any calculation
Instead of generating a separate launch script for all the calculation
classes provided by the plugin, we define a single one, that should be
able to serve almost all classes. Almost all cod-tools scripts, and by
extension the calculation plugins, require an input CIF file and some
command line parameters. The `aiida_codtools.cli.calculations.cod_tools`
script takes a code, a `CifData` node and will run the calculation for
which the code is configured. The command line parameters that one wants
to run can be passed as a single string.

To parse the single string into a dictionary, which is passed as an
input node to the calculation job, a new utility class is introduced in
`aiida_codtools.common.cli.CliParameters`. This container hosts a
representation of CLI parameters that can be passed to any cod-tools
script and provides methods to return them in string, list or dictionary
format.