👌 Add formula check to `CifCleanWorkChain`
Add a new step to the outline of the `CifCleanWorkChain` where the formula parsed
directly from the CIF file and the formula of the parsed structure are
compared. In case of a mismatch, or a failure of this check due to an unexpected
exception, the corresponding exit code is stored in the context and returned during the
`results` step in the outline.

Co-authored-by: Loris Ercole <loris.ercole@gmail.com>